+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## TBA
+
+- **docs:** fix outdated docs on `Policy.circuitBreaker` and unnecessary dashes in jsdoc comments (see [#22](https://github.com/connor4312/cockatiel/issues/22), [#23](https://github.com/connor4312/cockatiel/issues/23), [#24](https://github.com/connor4312/cockatiel/issues/24))
+
 ## 1.0.1 - 2020-06-22
 
 - **fix:** cockatiel not working in certain browser builds